#104 Start abstracting out some of the form metadata manipulation
