+++ src/ts/content/entities/bookForm/data/index.ts
@@ -1,14 +1,15 @@
-import {FormAreaElement, FormData} from "../types";
+import {FormAreaElement, FormData, FormMetaDataDecorator} from "../types";
 import {getFormElements} from "../util";
-import {ensureRolesInputCount, show} from "../ui";
+import {show} from "../ui";
 import {match} from "../../../../common/util/match";
-import {collections} from "./uniqueData/collections";
+import {collections, isCollectionsElement} from "./uniqueData/collections";
 import {physicalDescription} from "./uniqueData/physicalDescription";
 import {transformIncomingData} from "./transformData";
 import {matchFactoryFromDescriptors} from "./uniqueData/uniqueFormElement";
+import {ensureRolesInputCount, addRoleCountToMetaData} from "./uniqueData/roles";
 
 const FORM_META_DATA_KEY = "___metadata_";
-const ROLES_INPUT_COUNT_KEY = "___roles-count_";
+const PHYSICAL_DESCRIPTION_COUNTS_KEY = "___phys-desc-counts_";
 
 const matchFactory = matchFactoryFromDescriptors(collections, ...physicalDescription);
 
@@ -24,12 +25,11 @@ const internalIsFormDataElement = matchFactory("isFormData", (_formData, element
 // form metadata is data ABOUT the form. this could be extended in the future,
 // like ... for physical description
 const getFormMetadata = (document: Document): FormData => ({
-	[FORM_META_DATA_KEY]: {[ROLES_INPUT_COUNT_KEY]: getRolesInputCount(document)},
+	[FORM_META_DATA_KEY]: [addRoleCountToMetaData].reduce(
+		(metaData: Record<string, unknown>, callback: FormMetaDataDecorator) => callback(document, metaData),
+		{}
+	),
 });
-
-// We subtract one to omit the main author, who is not part of the roles section
-const getRolesInputCount = (document: Document): number =>
-	document.querySelectorAll("input.bookEditInput.bookPersonName").length - 1;
 
 // We can't change hidden elements because LibraryThing relies
 // on hidden form inputs to send additional, form-specific metadata
@@ -44,7 +44,7 @@ const getFormData = (_document = document) =>
 	}, getFormMetadata(_document));
 
 const insertFormData = (formData: FormData) => {
-	ensureRolesInputCount(formData?.[FORM_META_DATA_KEY]?.[ROLES_INPUT_COUNT_KEY] ?? 0);
+	ensureRolesInputCount(formData?.[FORM_META_DATA_KEY] ?? {});
 	getFormElements(document).forEach((element: any) => {
 		if (isFormDataElement(element)) {
 			const {value, checked} = transformIncomingData(extractFromFormData(formData, element), element);
@@ -62,7 +62,7 @@ const insertFormData = (formData: FormData) => {
 // to expose COLLECTIONS_ID_PREFIX
 const ensureVisible = (element: Element) =>
 	match(element)
-		.case(collections.predicate, show)
+		.case(isCollectionsElement, show)
 		.default((): void => undefined)
 		.yield();
